Calling find_finfo(3) on a file "a" identified by an absolute path would throw a segfault if another file sharing the same absolute path was previously recorded and the two were of different file types(file and folder).
That was because folders have a NULL hash and that would lead to a strcmp(path, NULL).

This commit solves the bug by adding an additional check so that the above scenario is ignored as it should be, since a file and a folder can't be the same file anyways.

Signed-off-by: Fotios Valasiadis <fvalasiad@gmail.com>