configure.ac: Add option to configure script for benchmarks.
As discussed in #205, this commit makes it so
that building the benchmark directory is optional
and disabled by default. I'd argue opt-in is superior
in our case since benchmarks are only relevant to the
developers of this project.

This allows us to add a as many dependencies to benchmark
as we possibly want or need, without minding the extra
dependencies that we add to the project.

We still have to face the fact that different benchmark
targets have different dependencies and perhaps we'd want
to provide options for each separate benchmark but I find
this too complicated.

Signed-off-by: Fotios Valasiadis <fvalasiad@gmail.com>