Handle PTRACE_EVENT_* events for child processes/threads.
While we so far support clone(2), fork(2), vfork(2), there
are other variants we don't. In my testing I found a case
where gcc utilized clone3(2), resulting in build-recorder
not recording the process creation step.

This commit follows a second approach utilizing the signals
ptrace(2) delivers to inform the tracer of a new process
or thread.

Signed-off-by: Fotios Valasiadis <fvalasiad@gmail.com>